==> 02_read_pdf.ipynb <==
import os
import pandas as pd
from PyPDF2 import PdfReader

def extract_text_from_pdfs(directory_path):
    data = {'filename': [], 'text': []}
    
    for filename in os.listdir(directory_path):
        if filename.endswith(".pdf"):
            file_path = os.path.join(directory_path, filename)
            
            # Extract text from PDF
            text = ""
            with open(file_path, "rb") as pdf_file:
                pdf_reader = PdfReader(pdf_file)
                for page in pdf_reader.pages:
                    text += page.extract_text() or ""
                    
            # Append filename and text to data
            data['filename'].append(filename)
            data['text'].append(text)
    
    # Create DataFrame
    df = pd.DataFrame(data)
    return df

# Usage
directory_path = "all_syllabi/"
df = extract_text_from_pdfs(directory_path)

import re
df['cleaned_text'] = df['text'].apply(lambda x: re.sub(r'\s+', ' ', x))
df['course name'] = df['filename'].str.split(' - ').str[0].str.split('_').str[0]
df['course department'] = df['course name'].str.split(' ').str[0]

df['course department'].value_counts()[:30]

ndf = pd.read_csv('name_mapping/ALL.csv')
ndf = ndf.drop_duplicates('code')
print(ndf['code'].nunique())
ndf[ndf['code'].duplicated()].sort_values('code')
ndf['code2'] = ndf['code'].str.split('/').str[0]
ndf

df = df.merge(ndf[['code2', 'Name']], left_on = 'course name', right_on = 'code2')
df = df[['filename',  'course department', 'course name','Name', 'text', 'cleaned_text']]
df

df1 = df[df['course department'].isin(['MATH','STAT','APCOMP','APMTH','BIOSTAT'])]
df1

df2 = df[~df['course department'].isin(['MATH','STAT','APCOMP','APMTH','BIOSTAT'])&(df['Name'].str.lower().str.contains('stat')|df['Name'].str.lower().str.contains('math')|df['Name'].str.lower().str.contains('quant') )]
df2

df3 = pd.concat([df1, df2])
df3

df3.to_csv('processed_data/computation_courses.csv', index = False)
